aes: move portable RSC loader into its own file
